Rearranged whole index (other pages will catch up. Need to redo the page creation class.)
Added notes to new folder. Added notes in index to add characters.

Partially updated the word 'game' to 'app' and the word 'app' to 'tool' in variable names, file names, and on pages.

Added snake to home page.

Created a page with a map that directs you from your location to Paris France. Plan to add this to the home page under the snake game. Not quite finished.